Target: move optimization_flags method to spack.build_environment
